Worked on review comments. Created acl-module.groovy file and moved deletePolicies().
