+++ builds/delete-serverless-service-build-pack/Jenkinsfile
@@ -22,6 +22,7 @@ import groovy.json.JsonSlurperClassic
 @Field def environmentMetadataLoader
 @Field def sonarModule
 @Field def lambdaEvents
+@Field def aclModule
 
 @Field def g_base_url = ''
 @Field def g_svc_admin_cred_ID = 'SVC_ADMIN'
@@ -247,7 +248,8 @@ node {
 
 	if (is_service_deletion) {
 		stage('Cleanup SCM') {
-			cleanup(configObj.service_id)
+			cleanup(repo_name)
+			aclModule.deletePolicies(configObj.service_id, auth_token, "${g_base_url}/jazz/acl/policies")
 			events.sendCompletedEvent('DELETE_PROJECT', 'deletion completed', context_map)
 		}
 	}
@@ -470,11 +472,9 @@ def updateSwaggerConfig() {
  * @param  repoName
  * @return
  */
-def cleanup(serviceId) {
-
+def cleanup(repoName) {
 	try {
-		def aclUrl = "${g_base_url}/jazz/acl/policies"
-		scmModule.updateAclPermission(serviceId, [], auth_token, aclUrl)
+    	scmModule.deleteProject(repoName)
 		send_status_email("COMPLETED")
 	} catch (ex) {
 		events.sendFailureEvent('DELETE_PROJECT', ex.getMessage(), context_map)
@@ -1192,10 +1192,12 @@ def loadBuildModules(buildModuleUrl){
 		sonarModule = load "sonar-module.groovy"
 		echo "Sonar module loaded successfully."
 
-    lambdaEvents = load "aws-lambda-events-module.groovy"
-    lambdaEvents.initialize(configLoader, utilModule)
-    echo "Lambda event module loaded successfully."
+		lambdaEvents = load "aws-lambda-events-module.groovy"
+		lambdaEvents.initialize(configLoader, utilModule)
+		echo "Lambda event module loaded successfully."
 
+    	aclModule = load "acl-module.groovy"
+		echo "ACL module loaded successfully."
 	}
 }
