+++ builds/jazz-build-module/scm-module.groovy
@@ -338,27 +338,4 @@ def getRepoCommitterInfo(commitHash) {
     }
 }
 
-def updateAclPermission(serviceId, policiesList, auth_token, aclUrl) {
-	try {
-		def body = JsonOutput.toJson([
-			serviceId: serviceId,
-			policies: policiesList
-		]);
-		def updatePermission = sh(script: "curl -X POST \
-				${aclUrl} \
-				-k -v -H \"Authorization: $auth_token\" \
-				-H \"Content-Type: application/json\" \
-				-d \'${body}\'", returnStdout: true).trim()
-		def responseJSON = parseJson(updatePermission)
-
-		if (responseJSON && responseJSON.success) {
-			echo "Successfully updated permissions for code and deploy."
-		} else {
-			echo "Something went wrong while updating permissions for code and deploy."
-		}
-	} catch(ex) {
-		echo "ex: $ex"
-	}
-}
-
 return this
